requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
black==21.5b1
chardet==4.0.0
click==8.0.1
discord-py-slash-command==1.1.2
discord.py==1.7.2
flake8==3.9.2
idna==3.1
isort==5.8.0
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pycodestyle==2.7.0
PyExecJS==1.5.1
pyflakes==2.3.1
pyparsing==2.4.7
pyTwistyScrambler==1.6
regex==2021.4.4
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
yarl==1.6.3
